+++ src/Controller/ContactLinkController.php
@@ -107,11 +107,16 @@ class ContactLinkController extends AbstractController
     #[Route(path: "/{id}", name: 'api_contact_link_delete', methods: ["DELETE"])]
     public function delete(ContactLink $contactLink, Request $request, EntityManagerInterface $entityManager): JsonResponse
     {
+        if (!$this->user) {
+            return new JsonResponse(['message' => 'User not authenticated'], JsonResponse::HTTP_UNAUTHORIZED);
+        }
+
         $data = $request->toArray();
         if (isset($data['force']) && $data['force'] === true) {
             $entityManager->remove($contactLink);
         } else {
             $contactLink->setStatus("off");
+            $contactLink->setUpdatedBy($this->user->getId());
             $entityManager->persist($contactLink);
         }
